more docs and cleanup
